pyproject.toml
[tool.poetry]
name = "tracking"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
aiofiles = "^0.8.0"
PyYAML = "^6.0"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
aiosqlite = "^0.17.0"
pytimeparse = "^1.1.8"
dependency-injector = "^4.39.1"

[tool.poetry.dev-dependencies]
pytest = "*"
black = "^22.3.0"
wemake-python-styleguide = "^0.16.1"
pycln = "^1.3.2"
mypy = "^0.950"
ipdb = "^0.13.9"
ipython = "^8.3.0"
pytest-asyncio = "^0.18.3"
syrupy = "^2.0.0"
pytest-vcr = "^1.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
